backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pymongo
python-dotenv
passlib[bcrypt]
pyjwt
pdfplumber
sentence-transformers
torch
numpy
scikit-learn